+++ packages/svelte/src/lib/providers/locale/index.ts
@@ -1,4 +1,5 @@
 export { default as LocaleProvider, type LocaleProviderProps } from './locale-provider.svelte'
 export { useCollator, type UseCollatorProps, type UseCollatorReturn } from './use-collator.svelte'
+export { useDateFormatter, type UseDateFormatterProps, type UseDateFormatterReturn } from './use-date-formatter.svelte'
 export { useFilter, type UseFilterProps, type UseFilterReturn } from './use-filter.svelte'
 export { useLocaleContext, type UseLocaleContext } from './use-locale-context'