+++ packages/vue/src/providers/locale/index.ts
@@ -1,4 +1,5 @@
 export { default as LocaleProvider, type LocaleProviderProps } from './locale-provider.vue'
 export { useCollator, type UseCollatorProps, type UseCollatorReturn } from './use-collator'
+export { useDateFormatter, type UseDateFormatterProps, type UseDateFormatterReturn } from './use-date-formatter'
 export { useFilter, type UseFilterProps, type UseFilterReturn } from './use-filter'
 export { DEFAULT_LOCALE, useLocaleContext, type LocaleContext } from './use-locale-context'